add variable data type interface, test to compare old vs new encoding
