+++ pkg/scale/decode.go
@@ -1,0 +1,1 @@
+package scale
